third_party/blink/web_tests/virtual/legacy-client-hints-no-fp-delegation/external/wpt/client-hints/critical-ch/request-count.https.window-expected.txt

This is a testharness.js-based test.
FAIL Critical-CH navigation assert_equals: expected "2" but got "1"
Harness: the test ran to completion.
